Implement error alert to all routes
