+++ src/components/Auth/Login.js
@@ -1,17 +1,14 @@
 /* eslint-disable operator-linebreak */
 import React, { useState } from 'react';
-import { useSelector, useDispatch } from 'react-redux';
+import { useDispatch } from 'react-redux';
 import { useNavigate } from 'react-router-dom';
 import bcrypt from 'bcryptjs';
 import { fetchUser } from '../../redux/features/user';
-import AlertComponent from '../Alert/Alert';
 import { triggerAlert } from '../../redux/features/alert';
 
 const salt = '$2a$10$CwTycUXWue0Thq9StjUM0u';
 
 const Login = () => {
-  const alert = useSelector((state) => state.alert);
-
   const [formData, setFormData] = useState({
     email: '',
     password: '',
@@ -60,41 +57,32 @@ const Login = () => {
     }
   };
   return (
-    <>
-      {alert.status && (
-        <AlertComponent
-          heading={alert.heading}
-          message={alert.message}
-          variant={alert.variant}
-        />
-      )}
-      <div className="login_container">
-        <h1>Welcome back</h1>
-        <form onSubmit={handleSubmit}>
-          <div className="login_field">
-            <p>Email:</p>
-            <input
-              type="email"
-              name="email"
-              placeholder="Enter your email"
-              onChange={handleInput}
-            />
-          </div>
-          <div className="login_field">
-            <p>Password:</p>
-            <input
-              type="password"
-              name="password"
-              placeholder="Enter your password"
-              onChange={handleInput}
-            />
-          </div>
-          <div className="login_field">
-            <button type="submit">Login</button>
-          </div>
-        </form>
-      </div>
-    </>
+    <div className="login_container">
+      <h1>Welcome back</h1>
+      <form onSubmit={handleSubmit}>
+        <div className="login_field">
+          <p>Email:</p>
+          <input
+            type="email"
+            name="email"
+            placeholder="Enter your email"
+            onChange={handleInput}
+          />
+        </div>
+        <div className="login_field">
+          <p>Password:</p>
+          <input
+            type="password"
+            name="password"
+            placeholder="Enter your password"
+            onChange={handleInput}
+          />
+        </div>
+        <div className="login_field">
+          <button type="submit">Login</button>
+        </div>
+      </form>
+    </div>
   );
 };
